{{cookiecutter.project_slug}}/backend/requirements.txt
autopep8==1.4
attrs==18.2.0
backcall==0.1.0
certifi==2018.8.24
chardet==3.0.4
decorator==4.3.0
Django==2.1.1
django-environ==0.4.5
django-extensions==2.1.2
django-redis==4.9.0
flake8==3.5.0
flake8-commas==2.0.0
flake8-mypy==17.8.0
gevent==1.3.6
greenlet==0.4.14
gunicorn==19.9.0
idna==2.7
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
mccabe==0.6.1
mypy==0.620
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
psycopg2-binary==2.7.5
ptyprocess==0.6.0
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pytz==2018.5
redis==2.10.6
requests==2.19.1
rope==0.11.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.23
wcwidth==0.1.7
Werkzeug==0.14.1
{% if cookiecutter.use_sentry == 'y' %}raven==6.9.0{% endif %}
{% if cookiecutter.api == 'REST' %}
djangorestframework==3.8.2
{% elif cookiecutter.api == 'GraphQL' %}
django-graphql-jwt==0.1.10
graphene==2.1.3
graphene-django==2.1.0
{% endif %}


